feat: add pdf ocr

+++ src/Hx.Abp.Attachment.Application.ArchAI.Contracts/Hx/Abp/Attachment/Application/ArchAI/Contracts/RecognizeCharacterDataRectangles.cs
@@ -2,10 +2,10 @@ namespace Hx.Abp.Attachment.Application.ArchAI.Contracts
 {
     public class RecognizeCharacterDataRectangles
     {
-        public int Angle { get; set; }
-        public int Height { get; set; }
-        public int Left { get; set; }
-        public int Top { get; set; }
-        public int Width { get; set; }
+        public long Angle { get; set; }
+        public long Height { get; set; }
+        public long Left { get; set; }
+        public long Top { get; set; }
+        public long Width { get; set; }
     }
 }
